added validation on movie form

+++ src/bioscoop_app/Service/UploadService.cs
@@ -77,7 +77,10 @@ namespace bioscoop_app.Service
         /// <param name="filename"></param>
         public void DeleteFile(string filename)
         {
-            File.Delete($"{StorageService.GetUploadPath()}{filename}");
+            if (File.Exists($"{StorageService.GetUploadPath()}{filename}"))
+            {
+                File.Delete($"{StorageService.GetUploadPath()}{filename}");
+            }
         }
 
         /// <summary>
